Moved NodeState to its own file
